m/space: Special case pure pose chains to be TRACKED
An xrt_pose that is upgraded to an xrt_space_relation is only VALID, not TRACKED.
apply_relation only introduces TRACKED flags if there is a TRACKED xrt_space_relation in the chain.
If only poses are in the chain, special case their relationship to be fully known/TRACKED.

Example: Locating LOCAL space in LOCAL space has once the local space offset pose
and once the inverted local space offset in the chain.

Partial follow up to !2121.